Dev v2

+++ functions/helper.sh
@@ -2,6 +2,36 @@
 
 # Helpers
 
+# usage function
+helper.usage() {
+   cat << USAGE
+
+   Usage: $(basename $0) [--num NUM] [--time TIME_STR] [--verbose] [--dry-run]
+
+   optional arguments:
+     -h, --help           show this help message and exit
+     -n, --num NUM        pass in a number
+     -t, --time T IME_STR  pass in a time string
+     -v, --verbose        increase the verbosity of the bash script
+     --dry-run            do a dry run, dont change any files
+
+USAGE
+}
+
+helper.welcome_message() {
+	local message
+
+	message="🆔 [@${message_new_chat_member_username[$id]:-null}]\n"
+    	message+="🗣 Olá *${message_new_chat_member_first_name[$id]}*"'!!\n\n'
+    	message+="Seja bem-vindo(a) ao *${message_chat_title[$id]}*.\n\n"
+    	message+='`Se precisar de ajuda ou informações sobre meus comandos, é só me chamar no privado.`'"[@$(ShellBot.username)]"
+
+	ShellBot.sendMessage --chat_id ${message_chat_id[$id]} \
+		--text "$(echo -e ${message})" --parse_mode markdown
+
+	return 0	
+}
+
 exitOnError() {
   # usage: exitOnError <output_message> [optional: code (defaul:exit code)]
   code=${2:-$?}
@@ -110,16 +140,24 @@ helper.remove_acento() {
 
 helper.get_api() {
 	
-  local tmp_folder
+  local tmp_folder current_version check_new_version
 
   tmp_folder=/tmp/$(helper.random)
+  check_new_version=$(curl -sS ${API_VERSION_RAW_URL} | grep VERSÃO | grep -o [0-9].*)
+  current_version=$(cat ${BASEDIR}/ShellBot.sh | grep VERSÃO | grep -o [0-9].*)
 
-  echo "Baixando versão mais atual da API ShellBot"
-  git clone ${API_GIT_URL} ${tmp_folder} > /dev/null
+  if [[ "${current_version}" != "${check_new_version}" ]]; then
 
-  echo "Habilitando a API"
-  cp ${tmp_folder}/ShellBot.sh ${BASEDIR}/
-  rm -fr ${tmp_folder}
+    echo "[INFO] ShellBot API - Getting the newest version '${check_new_version}'"
+    git clone ${API_GIT_URL} ${tmp_folder} > /dev/null
+
+    echo "Habilitando a API"
+    cp ${tmp_folder}/ShellBot.sh ${BASEDIR}/
+    rm -fr ${tmp_folder}
+
+  else
+    echo "[INFO] ShellBot API version is the same as in the local repository (version: '${current_version}')"
+  fi
 }
 
 helper.random() {
@@ -143,18 +181,4 @@ helper.random() {
 	fi
 
 	echo "${random_number}"
-}
-
-helper.welcome_message() {
-	local message
-
-	message="🆔 [@${message_new_chat_member_username[$id]:-null}]\n"
-    	message+="🗣 Olá *${message_new_chat_member_first_name[$id]}*"'!!\n\n'
-    	message+="Seja bem-vindo(a) ao *${message_chat_title[$id]}*.\n\n"
-    	message+='`Se precisar de ajuda ou informações sobre meus comandos, é só me chamar no privado.`'"[@$(ShellBot.username)]"
-
-	ShellBot.sendMessage --chat_id ${message_chat_id[$id]} \
-		--text "$(echo -e ${message})" --parse_mode markdown
-
-	return 0	
 }
